pyproject.toml
[project]
name = "ai-tea-party"
version = "2.1.0"
description = "AI Tea Party backend service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.5",
    "requests>=2.31.0",
    "litellm>=1.81.0",
    "aiosqlite>=0.22.1",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = ["archive/", "scripts/", "mockups/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501", "F841"]
